Remove multiple definitions for IntReader and IntInterpreter.
